backend/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python-headless
ultralytics
easyocr
python-multipart
